requirements.txt
opencv-python
pytesseract
tabula-py
